% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossover_basic_3d.R
\name{crossover_basic_3d}
\alias{crossover_basic_3d}
\title{crossover_basic_3d}
\usage{
crossover_basic_3d(self, population, par)
}
\arguments{
\item{self}{a gapackage::ga class}

\item{population}{an array representing a 3D population of genotypes}

\item{par}{a named list containing initial parameters passed to the gapackage::ga$new function}
}
\value{
an array representing a 2D population of genotypes
}
\description{
A function to perform a basic crossover algorithm on a 3D population.
It will split and merge a random selection of genotypes a predetermined number of times.
The function should never be run explicitly, only passed as a parameter to the add_operator()
function of the gapackage::ga class.
}
\details{
This algorithm will require no dependency functions but does require two parameters to be passed
to the gapackage::ga$new function. They are as follows:
\enumerate{
\item \strong{par$add_proportion} the number of genotypes be added as a percentage of the total population size.
\item \strong{par$population_size} the number of genotypes in the entire population.
}
}
