% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentiate.R
\name{differentiate}
\alias{differentiate}
\title{differentiate}
\usage{
differentiate(curve, time_interval)
}
\arguments{
\item{curve}{A 2D array with 2 columns where each row is a coordinate. The array is treated as ordered.}
}
\value{
A vector indicating the approximate gradient for each segment of the curve.
}
\description{
A function which will numerically differentiate an array of coordinates.
}
