% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_brachistochrone.R
\name{example_brachistochrone}
\alias{example_brachistochrone}
\title{example_brachistochrone}
\usage{
example_brachistochrone()
}
\description{
This function will run the brachistochrone example.
}
\details{
This function is used by the ga_example.
}
