% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_manhattan_distance.R
\name{example_manhattan_distance}
\alias{example_manhattan_distance}
\title{example_manhattan_distance}
\usage{
example_manhattan_distance()
}
\description{
This function will run the manhattan distance example.
}
\details{
This function is used by the ga_example.
}
