% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_non_euclidean_distance.R
\name{example_non_euclidean_distance}
\alias{example_non_euclidean_distance}
\title{example_non_euclidean_distance}
\usage{
example_non_euclidean_distance()
}
\description{
This function will run the non-euclidean distance example.
}
\details{
This function is used by the ga_example.
}
