% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_non_euclidean_distance_const.R
\name{example_non_euclidean_distance_const}
\alias{example_non_euclidean_distance_const}
\title{example_non_euclidean_distance_constS}
\usage{
example_non_euclidean_distance_const()
}
\description{
This function will run the non-euclidean distance example with a constant identity matrix of A.
}
\details{
This function is used by the ga_example.
}
