% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_non_euclidean_var.R
\name{example_non_euclidean_distance_var}
\alias{example_non_euclidean_distance_var}
\title{example_non_euclidean_distance_var}
\usage{
example_non_euclidean_distance_var()
}
\description{
This function will run the non-euclidean distance example with a variable matrix A.
}
\details{
This function is used by the ga_example.
}
