% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_example.R
\name{ga_example}
\alias{ga_example}
\title{ga_example}
\usage{
ga_example(name)
}
\arguments{
\item{name}{the name of the example}
}
\description{
This function will allow you to run a number of examples
}
\details{
available examples: 'euclidean'
}
