% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_settings.R
\name{get_settings}
\alias{get_settings}
\title{get_settings}
\usage{
get_settings(settings, label)
}
\arguments{
\item{settings}{a list of named lists containing information about what questions to ask the user}

\item{label}{a title to begin the questioning.}
}
\value{
a named list indicating the chosen values of the parameters.
}
\description{
This function will allow you to get user input for a variety of settings
}
\details{
This function is used by all the examples to prompt the user for information.
}
