% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_rand_uni_3d.R
\name{initial_rand_uni_3d}
\alias{initial_rand_uni_3d}
\title{initial_rand_uni_3d}
\usage{
initial_rand_uni_3d(self, par)
}
\arguments{
\item{self}{a gapackage::ga class}

\item{par}{a named list containing initial parameters passed to the gapackage::ga$new function}
}
\value{
an array representing a 3D population of genotypes
}
\description{
A function to generate an initial population as a 3D array. Each element will be
taken from a uniform distribution.
The function should never be run explicitly, only passed to the gapackage::ga$new() function
as the 'initial' parameter.
}
\details{
This algorithm will require no dependency functions but does require four parameters to be passed
to the gapackage::ga$new function. They are as follows:
\enumerate{
\item \strong{par$geno_length} an integer indicating the length of a single genotype (the number of rows in the population array)
\item \strong{par$population_size} the number of genotypes in the entire population.
\item \strong{par$initial_max} the 'max' parameter passed to the runif() function.
\item \strong{par$initial_min} the 'min' parameter passed to the runif() function.
}
}
