% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_cauchy.R
\name{mutation_cauchy}
\alias{mutation_cauchy}
\title{mutation_cauchy}
\usage{
mutation_cauchy(self, population, par)
}
\arguments{
\item{self}{a gapackage::ga class}

\item{population}{an array of any dimensions representing a population of genotypes.}

\item{par}{a named list containing initial parameters passed to the gapackage::ga$new function}
}
\value{
an array representing a population of genotypes with the same dimensions as the population parameter.
}
\description{
A function which will add a numeric value taken from a cauchy distribution to each
value of the population.
The function should never be run explicitly, only passed as a parameter to the add_operator()
function of the gapackage::ga class.
}
\details{
This algorithm will require no dependency functions but does require two parameters to be passed
to the gapackage::ga$new function. They are as follows:
\enumerate{
\item \strong{par$location} the location parameter of the cauchy distribution.
\item \strong{par$population_size} the scale parameter of the cauchy distribution.
}
}
