% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_basic_3d.R
\name{selection_basic_3d}
\alias{selection_basic_3d}
\title{selection_basic_3d}
\usage{
selection_basic_3d(self, population, par)
}
\arguments{
\item{self}{a gapackage::ga class}

\item{population}{an array representing a 3D population of genotypes.}

\item{par}{a named list containing initial parameters passed to the gapackage::ga$new function}
}
\value{
an array representing a population of genotypes with the same dimensions as the population parameter.
}
\description{
This function can be used to sort a 3D population array according to a given selection
criteria. This function will not run unless the required dependencies are provided.
The function should never be run explicitly, only passed as a parameter to the add_operator()
function of the gapackage::ga class.
}
\details{
This algorithm will require 1 direct dependency and 1 indirect dependency. They are listed below respectively:
\enumerate{
\item \strong{self$dependents$fitness_function} a function which can evaluate the fitnesses of a 3D population and return a vector of numeric values.
\item \strong{self$dependents$fitness_function_single} a function which can evalue the fitness of a single 2D genotype.
}

There are three parameters that need to be passed to the parameter arguement of the gapackage::ga$new function. They are as follows:
\enumerate{
\item \strong{par$add_proportion} the number of genotypes to be removed as a percentage of the total population size.
\item \strong{par$population_size} the number of genotypes in the entire population.
\item \strong{par$maximise} a boolean indicating whether the fitness function should be maximised or minimised.
}
}
