% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_add_start_end_3d.R
\name{utility_add_start_end_3d}
\alias{utility_add_start_end_3d}
\title{utility_add_start_end_3d}
\usage{
utility_add_start_end_3d(genotype, par)
}
\arguments{
\item{genotype}{a 2D array of numbers representing a genotype.}

\item{par}{a named list containing initial parameters passed to the gapackage::ga$new function.}
}
\value{
a 2D array representing a genotype.
}
\description{
This function will add two coordinates to the start and end of a 2D genotype (in a 3D context).
It is designed to be used as a dependency function for the utility_fitness_brachistochrone_3d() function.
}
\details{
this function requires no dependencies but takes two parameters which are as follows:
\enumerate{
\item \strong{par$A_y} THe total numeric height of the traversal space.
\item \strong{par$B_x} THe total numeric width of the traversal space.
}
}
