% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_crossover_bread_3d.R
\name{utility_crossover_bread_3d}
\alias{utility_crossover_bread_3d}
\title{utility_crossover_breed_3d}
\usage{
utility_crossover_bread_3d(gen_1, gen_2)
}
\arguments{
\item{gen_1}{a 2D array of numbers representing a genotype.}

\item{gen_2}{a 2D array of numbers representing a genotype.}
}
\value{
a vector containing the two resulting children.
}
\description{
This function will 'breed' two 2D genotypes using a basic crossover algorithm. It is designed
to be used as a dependancy function for the crossover_basic_3d() function.
}
\details{
this function requires no parameters or dependancies.
}
