% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_fitness_brachistochrone_2d.R
\name{utility_fitness_brachistochrone_2d}
\alias{utility_fitness_brachistochrone_2d}
\title{utility_fitness_brachistochrone_2d}
\usage{
utility_fitness_brachistochrone_2d(self, genotype, par)
}
\arguments{
\item{self}{a gapackage::ga class}

\item{genotype}{a 1D vector indicating the height of the curve at discrete intervals.}

\item{par}{a named list containing initial parameters passed to the gapackage::ga$new function}
}
\value{
The traversal time of the given path.
}
\description{
This function will calculate the time taken to traverse a path in 2D space. It
will use a 1D vector representing the height of points along the path at discrete intervals. It will
add two values, par$A_y to the start of the vector and 0 to the end of the vector before calculating the time.
}
\details{
There are no dependancies but there are three parameters that need to be passed to the parameter arguement of the gapackage::ga$new function. They are as follows:
\enumerate{
\item \strong{par$A_y} The numric height to be added to the beginning of the genotype vector.
\item \strong{par$g} the gravitational constant.
\item \strong{par$B_x} THe total numeric width of the traversal space.
}
}
