% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_fitness_brachistochrone_3d.R
\name{utility_fitness_brachistochrone_3d}
\alias{utility_fitness_brachistochrone_3d}
\title{utility_fitness_brachistochrone_3d}
\usage{
utility_fitness_brachistochrone_3d(self, genotype, par)
}
\arguments{
\item{self}{a gapackage::ga class}

\item{genotype}{a 2D array indicating the coordinates of points on the curve.}

\item{par}{a named list containing initial parameters passed to the gapackage::ga$new function}
}
\value{
The traversal time of the given path.
}
\description{
This function will calculate the time taken to traverse a path in 2D space. It
will use a 2D array representing the coordinates of points along the path. It will
add two values, (0, par$A_y) to the start of the array and (par$B_x, 0) to the end of the array before calculating the time.
}
\details{
There are no dependancies but there are three parameters that need to be passed to the parameter arguement of the gapackage::ga$new function. They are as follows:
\enumerate{
\item \strong{par$A_y} THe total numeric height of the traversal space.
\item \strong{par$g} the gravitational constant.
\item \strong{par$B_x} THe total numeric width of the traversal space.
}
}
