% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_fitness_manhattan_3d.R
\name{utility_fitness_manhattan_3d}
\alias{utility_fitness_manhattan_3d}
\title{utility_fitness_manhattan_3d}
\usage{
utility_fitness_manhattan_3d(self, genotype, par)
}
\arguments{
\item{self}{a gapackage::ga class}

\item{genotype}{an array representing a 2D genotypes}

\item{par}{a named list containing initial parameters passed to the gapackage::ga$new function}
}
\value{
a number representing the total distance.
}
\description{
A function to calculate the manhattan length of a curve in a 3D context.
}
\details{
This algorithm will require no dependency functions but does require two parameters to be passed
to the gapackage::ga$new function. They are as follows:
\enumerate{
\item \strong{par$manhattan_start} The starting coordinate to be added to the beginning of the curve.
\item \strong{par$manhattan_end} The final coordinate to be added to the end of the curve.
}
}
